Spark/requirements.txt
pyspark==3.2.0
kafka-python==2.0.2
cassandra-driver==3.25.0
